Fixed width of daylight savings offset to be Int32

+++ opcua/ua/uaprotocol_auto.py
@@ -3382,7 +3382,7 @@ class TimeZoneDataType(FrozenClass):
     """
 
     ua_types = [
-        ('Offset', 'Int16'),
+        ('Offset', 'Int32'),
         ('DaylightSavingInOffset', 'Boolean'),
                ]
